Renamed endpoints/boxscorescoreing to boxscorescoring.
